--- src/Collection/Step.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Wearesho\Bobra\Ubki\Collection;
-
-use Wearesho\Bobra\Ubki;
-
-/**
- * Class Step
- * @package Wearesho\Bobra\Ubki\Collection
- */
-class Step extends Ubki\BaseCollection
-{
-    public function type(): string
-    {
-        return Ubki\Element\Step::class;
-    }
-}
