+++ src/Data/Step/Collection.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Wearesho\Bobra\Ubki\Data\Step;
+
+use Wearesho\Bobra\Ubki\BaseCollection;
+
+/**
+ * Class Entity
+ * @package Wearesho\Bobra\Ubki\Data\Step
+ */
+class Collection extends BaseCollection
+{
+    public function type(): string
+    {
+        return Entity::class;
+    }
+}
