+++ tests/Data/Step/EntityTest.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace Wearesho\Bobra\Ubki\Tests\Element;
+namespace Wearesho\Bobra\Ubki\Tests\Data\Step;
 
 use Carbon\Carbon;
 
 use Wearesho\Bobra\Ubki;
 
 /**
- * Class StepTest
+ * Class EntityTest
  * @internal
- * @package Wearesho\Bobra\Ubki\Tests\Element
+ * @package Wearesho\Bobra\Ubki\Tests\Data\Step
  */
-class StepTest extends Ubki\Tests\Extend\ElementTestCase
+class EntityTest extends Ubki\Tests\Extend\ElementTestCase
 {
     protected const TAG = 'step';
 
-    /** @var Ubki\Element\Step */
+    /** @var Ubki\Data\Step\Entity */
     protected $element;
 
     /** @var Carbon */
@@ -54,13 +54,13 @@ class StepTest extends Ubki\Tests\Extend\ElementTestCase
     public function testInvalidDate(): void
     {
         $this->start->addMinutes(10);
-        $this->element = new Ubki\Element\Step('test', $this->start, $this->end);
+        $this->element = new Ubki\Data\Step\Entity('test', $this->start, $this->end);
     }
 
     protected function setUp(): void
     {
         $this->start = Carbon::now();
         $this->end = Carbon::instance($this->start)->addMinute();
-        $this->element = new Ubki\Element\Step('build report', $this->start, $this->end);
+        $this->element = new Ubki\Data\Step\Entity('build report', $this->start, $this->end);
     }
 }
